refactored menus

+++ NumberLists/NumberList.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 
 namespace NumberLists
 {
-    public class NumberList : List<int>
+    public class NumberList : List<long>
     {
         //display numbers in one line with spaces between
         public void WriteListWithSpacesAndNewLine()
         {
             Console.WriteLine();
-            foreach (int number in this)
+            foreach (long number in this)
             {
                 Console.Write($"{number} ");
             }
